Get rid of helm-object-model.
