+++ Tasks/KubernetesManifestV0/task.json
@@ -15,7 +15,7 @@
     "version": {
         "Major": 0,
         "Minor": 151,
-        "Patch": 6
+        "Patch": 7
     },
     "demands": [],
     "groups": [],
